# swagger.yml
openapi: 3.0.0
info:
  title: Hillel Task App API
  version: 0.1.0
servers:
  - url: 'http://localhost:8000/api'
    description: Local server
  - url: 'https://tomcat.ownerofglory.com/hillel-task-app/api'
    description: Production server
...
paths:
  /boards:
    get:
      tags:
        - Boards
      summary: Returns a list of boards
      responses:
        200:
          description: A list of boards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
    post:
      tags:
        - Boards
      summary: Create a board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardWithoutId'
      responses:
        201:
          description: Created board item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
  /boards/{id}:
    get:
      tags:
        - Boards
      summary: Returns a specific board by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Board ID
          schema:
            type: integer
      responses:
        200:
          description: A board item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
    put:
      tags:
        - Boards
      summary: Updates a board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Board'
      responses:
        200:
          description: Updated board item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
    delete:
      tags:
        - Boards
      summary: Deletes a board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Deleted board item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
...
components:
  schemas:
    Board:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Home"
    BoardWithoutId:
      type: object
      properties:
        name:
          type: string
          example: "Home"
...